+++ route_editor/src/Settings.cpp
@@ -35,10 +35,10 @@ settings_t::settings_t()
     , fovy(64.0f)
     , fovy_min(2.0f)
     , fovy_max(100.0f)
-    , camera_initial_height(0.0)
-    , camera_move_speed(100.0)
-    , camera_rotate_speed(3.0)
-    , camera_zoom_power(250.0)
+    , camera_initial_height(0.0f)
+    , camera_move_speed(100.0f)
+    , camera_rotate_speed(3.0f)
+    , camera_zoom_power(250.0f)
     , gizmo_arrow_length(5.0f)
     , gizmo_arrow_thickness(0.1f)
     , gizmo_arrow_x_color(1.0f, 0.0f, 0.0f)
@@ -96,10 +96,10 @@ void settings_t::read(const std::string& cfg_path)
     cfg.getFloat(section, "FovY", fovy);
     cfg.getFloat(section, "FovYMin", fovy_min);
     cfg.getFloat(section, "FovYMax", fovy_max);
-    cfg.getDouble(section, "InitialHeight", camera_initial_height);
-    cfg.getDouble(section, "MoveSpeed", camera_move_speed);
-    cfg.getDouble(section, "RotateSpeed", camera_rotate_speed);
-    cfg.getDouble(section, "ZoomPower", camera_zoom_power);
+    cfg.getFloat(section, "InitialHeight", camera_initial_height);
+    cfg.getFloat(section, "MoveSpeed", camera_move_speed);
+    cfg.getFloat(section, "RotateSpeed", camera_rotate_speed);
+    cfg.getFloat(section, "ZoomPower", camera_zoom_power);
 
     section = "Gizmo";
